Cascaded two 74HC165s for extra settings buttons

+++ keyRegister.hpp
@@ -8,5 +8,5 @@
 #define SPKR 9
 
 void initKeyRegister();
-byte keyShiftIn();
+int keyShiftIn();
